+++ game/ui/PrintReports.java
@@ -25,6 +25,7 @@ public class PrintReports {
         }
         MyUtils.SteppedPrinting("GOVERNMENT TYPE: " + EnumToString.govtType(county.governmentType), 20);   
         MyUtils.SteppedPrinting("REAL GDP: " + MyUtils.formatNumber(county.realGdp)  + " Naira", 20);
+        MyUtils.SteppedPrinting("NOMINAL GDP: " + MyUtils.formatNumber(county.gdp)  + " Naira", 20);
         MyUtils.SteppedPrinting("POPULATION: " + MyUtils.formatNumber(county.population)  + " People", 20);
         MyUtils.SteppedPrinting("POSITION: " + MyUtils.Ordinalize(county.position) + " position out of 37 the states (Including Abuja)", 20);
         MyUtils.SteppedPrinting("PERFORMANCE SCORE: " + String.format("%.2f", county.rankingScore), 20);
@@ -38,7 +39,9 @@ public class PrintReports {
         MyUtils.SteppedPrinting("STATE MONTHLY REVENUE: " + MyUtils.formatNumber(county.monthlyRevenue) + " Naira", 20);
         MyUtils.SteppedPrinting("STATE POLICY SPEND: " + MyUtils.formatNumber(county.policySpend) + " Naira", 20);
         MyUtils.SteppedPrinting("BAIL OUT COUNT: " + county.crashCount, 20);
+        MyUtils.SteppedPrinting("PRICE OF OIL PER BARREL: " + MyUtils.formatNumber(fed.oilPriceIndex)  + " Naira", 20);  
         MyUtils.SteppedPrinting("OPERATING CASH: " + MyUtils.formatNumber(county.cash)  + " Naira", 20);  
+        MyUtils.SteppedPrinting("INFLATION RATE: " + String.format("%.2f",county.inflationRate * 100)  + "%", 20);  
         System.out.println(" ");  
     }
 
@@ -60,7 +63,7 @@ public class PrintReports {
         MyUtils.SteppedPrinting("NATIONAL GDP: " + MyUtils.formatNumber(fed.nationalGDP) + " Naira", 25);
         MyUtils.SteppedPrinting("NATIONAL POPULATION: " + MyUtils.formatNumber(fed.nationalPopulation) + " People", 25);
         MyUtils.SteppedPrinting("FEDERAL RESERVE VALUE " + MyUtils.formatNumber(fed.federalReserve) + " Naira", 25);
-        MyUtils.SteppedPrinting("FEDERATION ACCOUNT BALANCE: " + MyUtils.formatNumber(fed.allocationPool) + " Naira", 25);  
+        MyUtils.SteppedPrinting("FEDERATION ALLOCATION THIS MONTH: " + MyUtils.formatNumber(fed.allocationPool) + " Naira", 25);  
         MyUtils.SteppedPrinting("PRICE OF OIL PER BARREL: " + MyUtils.formatNumber(fed.oilPriceIndex) + " Naira", 25);  
         System.out.println(" ");  
     }